Fixes for  the following issues: 1. Fix taskqueues drain/free to fix panic seen when interface is being    bought down and in parallel asynchronous link events happening.
2. Fix bxe_ifmedia_status()

Submitted by:Vaishali.Kulkarni@cavium.com and Anand.Khoje@cavium.com
MFC after:5 days


git-svn-id: svn+ssh://svn.freebsd.org/base/head@336438 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ sys/sparc64/sparc64/spitfire.c
@@ -1,4 +1,6 @@
 /*-
+ * SPDX-License-Identifier: BSD-2-Clause-FreeBSD
+ *
  * Copyright (c) 2003 Jake Burkholder.
  * All rights reserved.
  *
@@ -130,7 +132,7 @@ spitfire_icache_page_inval(vm_paddr_t pa)
 		    : "=r" (tag) : "r" (addr), "n" (ASI_ICACHE_TAG));
 		if (((tag >> IC_VALID_SHIFT) & IC_VALID_MASK) == 0)
 			continue;
-		tag &= IC_TAG_MASK << IC_TAG_SHIFT;
+		tag &= (u_long)IC_TAG_MASK << IC_TAG_SHIFT;
 		if (tag == target) {
 			PMAP_STATS_INC(spitfire_icache_npage_inval_match);
 			stxa_sync(addr, ASI_ICACHE_TAG, tag);
